# types: use (*math/big.Int).BitLen() == 0 to check if value is 0