--- a/types/uint.go
+++ b/types/uint.go
@@ -161,7 +161,7 @@ func (u *Uint) MarshalTo(data []byte) (n int, err error) {
 	if u.i == nil {
 		u.i = new(big.Int)
 	}
-	if len(u.i.Bytes()) == 0 {
+	if u.i.BitLen() == 0 { // The value 0
 		copy(data, []byte{0x30})
 		return 1, nil
 	}