@odeke-em 2021-02-15
>I've requested just the addition of a test case; looks good otherwise (and please feel free to dismiss my review once the test case is added)

Thank you for the review @alessio! I've added the tests, please take a look.